move _find_ml_task from cross_validate.py to its own module
